+++ utils/helpers.js
@@ -1,0 +1,5 @@
+function checkNan(str) {
+  return isNaN(str) || (str === null) ? '' : str.toString();
+}
+
+export {checkNan};
